Testing error of editor

+++ app/src/main/java/com/mediaghor/rainbowtools/Activities/MainActivity.java
@@ -1,7 +1,10 @@
 package com.mediaghor.rainbowtools.Activities;
 
 import android.annotation.SuppressLint;
+import android.os.AsyncTask;
 import android.os.Bundle;
+import android.util.Log;
+import android.view.View;
 
 import androidx.activity.EdgeToEdge;
 import androidx.appcompat.app.AppCompatActivity;
@@ -12,74 +15,58 @@ import com.google.android.material.tabs.TabLayout;
 import com.mediaghor.rainbowtools.Adapter.ViewPagerCategoryAdapter;
 import com.mediaghor.rainbowtools.R;
 
+import org.apache.xmlbeans.impl.xb.xsdschema.Public;
+
 import java.io.File;
 
 public class MainActivity extends AppCompatActivity {
-    TabLayout tab;
-    ViewPager viewPager;
+    private TabLayout tab;
+    private ViewPager viewPager;
+    View toolbar;
 
     @SuppressLint({"MissingInflatedId", "WrongViewCast"})
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         setTheme(R.style.MainActivity);
-
-        Toolbar toolbar;
         super.onCreate(savedInstanceState);
         EdgeToEdge.enable(this);
         setContentView(R.layout.activity_main);
-        toolbar = findViewById(R.id.tool_bar_global);
-        setSupportActionBar(toolbar);
 
-        //Set The Tab Layout And Viewpager For Categories
+        // Initialize Toolbar
+        toolbar = findViewById(R.id.layout_toolbar_included);
+
+        // Initialize TabLayout and ViewPager
         tab = findViewById(R.id.tabLayoutId);
         viewPager = findViewById(R.id.viewPagerId);
         ViewPagerCategoryAdapter adapter = new ViewPagerCategoryAdapter(getSupportFragmentManager());
         viewPager.setAdapter(adapter);
         tab.setupWithViewPager(viewPager);
 
+        // Run cache cleanup asynchronously
+        new DeleteCacheTask().execute(getCacheDir());
+    }
 
 
-        try {
-            File cacheDir = getCacheDir();
-            long cacheSize = getFolderSize(cacheDir);
-
-            if (cacheSize > 2000 * 1024) { // 200MB
+    // AsyncTask for cache deletion (Runs in Background)
+    private static class DeleteCacheTask extends AsyncTask<File, Void, Void> {
+        @Override
+        protected Void doInBackground(File... files) {
+            File cacheDir = files[0];
+            if (cacheDir != null) {
                 deleteCache(cacheDir);
             }
-        } catch (Exception e) {
-            e.printStackTrace();
+            return null;
         }
-    }
 
-
-    // Function to get the total cache size
-    private long getFolderSize(File folder) {
-        long size = 0;
-        try {
-            for (File file : folder.listFiles()) {
-                if (file.isDirectory()) {
-                    size += getFolderSize(file);
-                } else {
-                    size += file.length();
+        private void deleteCache(File folder) {
+            if (folder != null && folder.exists()) {
+                for (File file : folder.listFiles()) {
+                    if (file.isDirectory()) {
+                        deleteCache(file);
+                    }
+                    file.delete();
                 }
             }
-        } catch (Exception e) {
-            e.printStackTrace();
-        }
-        return size;
-    }
-
-    // Function to clear cache
-    private void deleteCache(File folder) {
-        try {
-            for (File file : folder.listFiles()) {
-                if (file.isDirectory()) {
-                    deleteCache(file);
-                }
-                file.delete();
-            }
-        } catch (Exception e) {
-            e.printStackTrace();
         }
     }
 }
